Added SaveBoxes and DamagePlanes
Added a SaveBox prefab that uses persistent data for the player location
to load in the player position and save it out to a .dat file. Also made
a script to damage the player with some knockback (to avoid repetitive
damaging). The amount of damage and knockback is determined via
variables.