Push Box difficulties
Push Box mechanic is not behaving as expected in gameplay compared to
editor use. Need to consult with Sean on this and see if we can find a
solution. In the meantime, I've created a 'simplified' box-push mechanic
that is just a rigidbody cube that resets position if it leaves a given
area.