Refined Line of Collectables
Line of collectables is now faster to adjust, more responsive, and from
what I can tell, doesn't have the CTRL+Z issue any longer. The entire
system also has a sphere 'handle', allowing a line of coins to be moved
as a group without changing the line shape, and it also helps declutter
the hierarchy.